Convert to python3

+++ bingads/v11/bulk/entities/target_criterions/bulk_campaign_negative_location_criterion.py
@@ -53,7 +53,7 @@ class BulkCampaignNegativeLocationCriterion(_SingleRecordBulkEntity):
         _SimpleBulkMapping(
             _StringTable.Target,
             field_to_csv=lambda c: field_to_csv_LocationTarget(c.negative_campaign_criterion),
-            csv_to_field=lambda c, v: csv_to_field_LocationTarget(c.negative_campaign_criterion, long(v) if v else None)
+            csv_to_field=lambda c, v: csv_to_field_LocationTarget(c.negative_campaign_criterion, int(v) if v else None)
         ),
         _SimpleBulkMapping(
             _StringTable.SubType,
