+++ bingads/v11/bulk/entities/target_criterions/bulk_campaign_radius_criterion.py
@@ -63,7 +63,7 @@ class BulkCampaignRadiusCriterion(_SingleRecordBulkEntity):
         _SimpleBulkMapping(
             _StringTable.Radius,
             field_to_csv=lambda c: field_to_csv_Radius(c.biddable_campaign_criterion),
-            csv_to_field=lambda c, v: csv_to_field_Radius(c.biddable_campaign_criterion, long(v) if v else None)
+            csv_to_field=lambda c, v: csv_to_field_Radius(c.biddable_campaign_criterion, int(v) if v else None)
         ),
         _SimpleBulkMapping(
             _StringTable.Unit,
